cloud-function/requirements.txt
google-cloud-pubsub==2.19.0
google-cloud-bigquery==3.20.1
